Transactions credit and debit return Money or null
The model was returning the minor units.
The transactions model also had a `currency` property, which has been
disabled. It clashed with the database column "currency". The
`setCurrency()` method will probably also be removed.
A Laravel morphing relationship is also added for the transactions to
the "related object". If using that instead of the legacy methods,
then class aliases can be used instead of FQDNs.